Refactor BTable class
