+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BTable.java
@@ -47,9 +47,6 @@ public class BTable implements BRefType<Object>, BCollection {
     protected BStructType constraintType;
     private BStringArray primaryKeys;
     private BStringArray indices;
-    private boolean isInMemoryTable;
-    // This attribute is relevant only for tables created through SQL connector select action
-    private boolean loadSQLTableToMemory;
     private boolean tableClosed;
 
     public BTable() {
@@ -59,18 +56,6 @@ public class BTable implements BRefType<Object>, BCollection {
         this.hasNextVal = false;
         this.tableName = null;
         this.constraintType = null;
-        this.isInMemoryTable = false;
-    }
-
-    public BTable(DataIterator dataIterator, boolean loadSQLTableToMemory) {
-        this.iterator = dataIterator;
-        this.nextPrefetched = false;
-        this.hasNextVal = false;
-        this.tableProvider = null;
-        this.tableName = null;
-        this.constraintType = null;
-        this.isInMemoryTable = false;
-        this.loadSQLTableToMemory = loadSQLTableToMemory;
     }
 
     public BTable(String tableName, BStructType constraintType) {
@@ -78,7 +63,6 @@ public class BTable implements BRefType<Object>, BCollection {
         this.hasNextVal = false;
         this.tableProvider = null;
         this.tableName = tableName;
-        this.isInMemoryTable = false;
         this.constraintType = constraintType;
     }
 
@@ -91,7 +75,6 @@ public class BTable implements BRefType<Object>, BCollection {
             this.tableName = tableProvider.createTable(fromTable.tableName, query, constraintType, params);
         }
         this.constraintType = constraintType;
-        this.isInMemoryTable = true;
     }
 
     public BTable(BType type, BStruct configStruct) {
@@ -113,7 +96,6 @@ public class BTable implements BRefType<Object>, BCollection {
         this.constraintType = (BStructType) constrainedType;
         this.primaryKeys = primaryKeys;
         this.indices = indexColumns;
-        this.isInMemoryTable = true;
         //Insert initial data
         if (data != null) {
             insertInitialData(data);
@@ -127,9 +109,6 @@ public class BTable implements BRefType<Object>, BCollection {
 
     @Override
     public String stringValue() {
-        if (!this.isInMemoryTable) {
-            return "";
-        }
         String constraint = constraintType != null ? "<" + constraintType.toString() + ">" : "";
         StringBuilder tableWrapper = new StringBuilder("table" + constraint + " ");
         StringJoiner tableContent = new StringJoiner(", ", "{", "}");
@@ -184,7 +163,7 @@ public class BTable implements BRefType<Object>, BCollection {
         return hasNextVal;
     }
 
-    public void next() {
+    public void moveToNext() {
         if (tableClosed) {
             throw new BallerinaException("Trying to perform an operation over a closed table");
         }
@@ -206,25 +185,16 @@ public class BTable implements BRefType<Object>, BCollection {
     }
 
     public void reset(boolean isInTransaction) {
-        if (loadSQLTableToMemory) {
-            iterator.reset(false);
-        } else if (isInMemoryTable) {
-            if (iterator != null) {
-                iterator.reset(false);
-                iterator = null;
-            }
-        } else {
-            // Table from SQL connector, loadSQLTableToMemory=false
-            if (iterator != null) {
-                iterator.reset(isInTransaction);
-            }
+        if (iterator != null) {
+            iterator.reset(isInTransaction);
+            iterator = null;
         }
         resetIterationHelperAttributes();
     }
 
     public BStruct getNext() {
         // Make next row the current row
-        next();
+        moveToNext();
         // Create BStruct from current row
         return iterator.generateNext();
     }
@@ -245,9 +215,6 @@ public class BTable implements BRefType<Object>, BCollection {
     }
 
     public void addData(BStruct data, Context context) {
-        if (!this.isInMemoryTable) {
-            throw new BallerinaException("data cannot be added to a table returned from a database");
-        }
         if (data.getType() != this.constraintType) {
             throw new BallerinaException("incompatible types: struct of type:" + data.getType().getName()
                     + " cannot be added to a table with type:" + this.constraintType.getName());
@@ -268,9 +235,6 @@ public class BTable implements BRefType<Object>, BCollection {
      */
     public void performRemoveOperation(Context context, BFunctionPointer lambdaFunction) {
         try {
-            if (!this.isInMemoryTable) {
-                throw new BallerinaException("data cannot be deleted from a table returned from a database");
-            }
             int deletedCount = 0;
             while (this.hasNext(false)) {
                 BStruct data = this.getNext();
@@ -278,7 +242,7 @@ public class BTable implements BRefType<Object>, BCollection {
                 BValue[] returns = BLangFunctions.invokeCallable(lambdaFunction.value().getFunctionInfo(), args);
                 if (((BBoolean) returns[0]).booleanValue()) {
                     ++deletedCount;
-                    this.removeData(data);
+                    tableProvider.deleteData(tableName, data);
                 }
             }
             context.setReturnValues(new BInteger(deletedCount));
@@ -286,10 +250,6 @@ public class BTable implements BRefType<Object>, BCollection {
         } catch (Throwable e) {
             context.setReturnValues(TableUtils.createTableOperationError(context, e));
         }
-    }
-
-    private void removeData(BStruct data) {
-        tableProvider.deleteData(tableName, data);
     }
 
     public String getString(int columnIndex) {
@@ -344,7 +304,7 @@ public class BTable implements BRefType<Object>, BCollection {
     }
 
     protected boolean isIteratorGenerationConditionMet() {
-        return this.isInMemoryTable && this.iterator == null;
+        return this.iterator == null;
     }
 
     protected void resetIterationHelperAttributes() {
@@ -354,12 +314,10 @@ public class BTable implements BRefType<Object>, BCollection {
 
     @Override
     protected void finalize() {
-        if (this.isInMemoryTable) {
-            if (this.iterator != null) {
-                this.iterator.close(false);
-            }
-            tableProvider.dropTable(this.tableName);
+        if (this.iterator != null) {
+            this.iterator.close(false);
         }
+        tableProvider.dropTable(this.tableName);
     }
 
     private void insertInitialData(BRefValueArray data) {
@@ -370,10 +328,6 @@ public class BTable implements BRefType<Object>, BCollection {
             }
             addData((BStruct) data.get(i));
         }
-    }
-
-    protected boolean iteratorResetRequired() {
-        return isInMemoryTable;
     }
 
     /**
